Ready For v1.00 Release
Mod is not ready for full v1.00 release as soon as the next live build is out where required PR alterations to core code-base files necessary for this mod to function. Also of note, this currently has compatibility issues with the "Inventory Filter" due to conflicting overrides for the trade window. Load this mod after that one for best results.